Moved Animation and Input to Core namespace
