+++ Assets/Scripts/Core/Animation/AnimatorController.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Player.PlayerAnimation
+namespace Core.Animation
 {
     public abstract class AnimatorController : MonoBehaviour
     {
