+++ Assets/Scripts/Core/Animation/UnityAnimatorController.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Player.PlayerAnimation
+namespace Core.Animation
 {
     [RequireComponent(typeof(Animator))]
     public class UnityAnimatorController : AnimatorController
