+++ Assets/Scripts/Core/GameLevelInitializer.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Core.InputReader;
 using Player;
 using UnityEngine;
