+++ Assets/Scripts/Core/InputReader.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 72cd4ca17cdc491fa33aacd978142d59
+timeCreated: 1680082374
