+++ Assets/Scripts/Core/InputReader/GameUIInputView.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.UI;
 
-namespace Player
+namespace Core.InputReader
 {
     public class GameUIInputView : MonoBehaviour, IEntityInputSource
     {
