+++ Assets/Scripts/Core/InputReader/IEntityInputSource.cs
@@ -1,4 +1,4 @@
-﻿namespace Player
+﻿namespace Core.InputReader
 {
     public interface IEntityInputSource
     {
