+++ Assets/Scripts/Player/PlayerBrain.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Core.InputReader;
 using UnityEngine;
 
 namespace Player
